automatically backed up by learn

+++ lib/bachelor.rb
@@ -65,7 +65,7 @@ def get_average_age_for_season(data, season)
    age_season_array.each do |age|
     age_sum = age_sum + age
    end
-average_age = age_sum / age_season_array.length.to_f   
+average_age = age_sum / age_season_array.length.to_i   
  
 end
